#76: make contact API validators consistent with others
