+++ API Server/publicAPIs/contact.js
@@ -4,6 +4,19 @@
 // https://opensource.org/licenses/MIT
 
 const nodemailer = require("nodemailer");
+const { body } = require("express-validator");
+
+const CONTACT_TYPES = ["Help", "Bug Report", "Suggestion", "Feedback", "General"];
+
+const validators = {
+    schema: [
+        body("type").isIn(CONTACT_TYPES),
+        body("subject").trim().escape(),
+        body("name").trim().escape(),
+        body("email").isEmail().normalizeEmail(),
+        body("message").trim().escape(),
+    ],
+};
 
 const post = (req, res) => {
     const { EMAIL_HOST, EMAIL_PORT, EMAIL_USERNAME, EMAIL_PASSWORD } = process.env;
@@ -33,7 +46,7 @@ const post = (req, res) => {
     });
     transport.sendMail(emailData, (err) => {
         if (err) {
-            res.status(500).send(err.toString());
+            res.status(500).send(err.message);
             return;
         }
         res.status(200).send("contact sent sucessfully");
@@ -42,4 +55,5 @@ const post = (req, res) => {
 
 module.exports = {
     post,
+    validators,
 };
